Completed shaders manager; Fixed all compilation and linker issues with DX11 renderer

+++ Code/Common/RenderModule/ITexture.h
@@ -17,7 +17,7 @@ enum class TextureType : uint8
 	Total
 };
 
-enum class TextureFormat : uint8
+enum class TextureFormat : uint16
 {
 	Unknown = 0,
